# storage/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LouvorJA API",
        "description": "API do LouvorJA - Gerenciador de Hinários e Músicas para a Igreja Adventista do Sétimo Dia. Fornece endpoints públicos para consulta de músicas, letras, arquivos e álbuns, além de endpoints administrativos protegidos por autenticação JWT.",
        "contact": {
            "name": "LouvorJA",
            "email": "contato@louvorja.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.louvorja.com.br",
            "description": "Produção"
        },
        {
            "url": "http://localhost:8000",
            "description": "Desenvolvimento"
        }
    ],
    "paths": {
...
        "/{lang}/collections/online": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Coleção de conteúdo online (público)",
                "description": "Retorna canais, playlists e vídeos online estruturados, com cache de 10 minutos",
                "operationId": "92b75d20b0de544e0d86944b5a507440",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "Código do idioma",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "pt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conteúdo online",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação JWT via header Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}